Separate fake protocols for iot and smart
